requirements.txt
Flask==3.1.1
Flask-Migrate==4.0.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
python-dotenv==1.0.1
Werkzeug==3.1.1